+++ iraspa/iraspagui/atomtreeviewdropcopycommand.h
@@ -34,7 +34,7 @@ class AtomTreeViewDropCopyCommand : public QUndoCommand
 {
 public:
   AtomTreeViewDropCopyCommand(MainWindow *mainWindow, AtomTreeViewModel *model, std::shared_ptr<iRASPAStructure> iraspaStructure,
-                              std::vector<std::tuple<std::shared_ptr<SKAtomTreeNode>, std::shared_ptr<SKAtomTreeNode>, int>> moves, QUndoCommand *undoParent = nullptr);
+                              std::vector<std::tuple<std::shared_ptr<SKAtomTreeNode>, std::shared_ptr<SKAtomTreeNode>, size_t>> moves, QUndoCommand *undoParent = nullptr);
   void redo() override final;
   void undo() override final;
 private:
@@ -42,6 +42,6 @@ private:
   AtomTreeViewModel *_model;
   std::shared_ptr<iRASPAStructure> _iraspaStructure;
   std::shared_ptr<Structure> _structure;
-  std::vector<std::tuple<std::shared_ptr<SKAtomTreeNode>, std::shared_ptr<SKAtomTreeNode>, int>> _moves;
-  std::vector<std::tuple<std::shared_ptr<SKAtomTreeNode>, std::shared_ptr<SKAtomTreeNode>, int>> _reverseMoves;
+  std::vector<std::tuple<std::shared_ptr<SKAtomTreeNode>, std::shared_ptr<SKAtomTreeNode>, size_t>> _moves;
+  std::vector<std::tuple<std::shared_ptr<SKAtomTreeNode>, std::shared_ptr<SKAtomTreeNode>, size_t>> _reverseMoves;
 };
